+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 ### Fixed
 - https://github.com/pulumi/pulumi-kubernetes/pull/4181 implements a fix where updating a persistent volume claim to a storage class with a bind mode of WaitForFirstConsumer no longer results in a timeout.
 
+### Added
+
+- https://github.com/pulumi/pulumi-kubernetes/pull/3661 - support for OCI login inside the provider process. Providing a username/password to `Chart v4` and `Release v3` will now login to private OCI
+  registries as expected. 
+  (https://github.com/pulumi/pulumi-kubernetes/issues/2911 and https://github.com/pulumi/pulumi-kubernetes/issues/1914)
+
 ## 4.26.0 (February 19, 2026)
 
 ### Fixed